chore: consolidate empty code change entries for clarity
